Revert to standard interpreter for running the examples tests
Using buildout-generated `bin/py` to run the examples with all required packages
in the path seemed to be convenient but breaks when the virtualenv is in a
different path than the nagiosplugin checkout. This happens to be the case on
our build server. Instead use the standard interpreter (sys.executable) again
and export sys.path to the spawned child process so that it finds all packages
that are accessible from the test runner.

Re #12297